Translation edit
I figured "napkin" would go over a lot of players heads in regards to what the teacher is saying here. "sanitary items" allows the player to know that she is probably refering to sanitary pads/tampons without her having to directly say it as she would probably want to avoid embarassing the girls too much infront of a boy.